# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mahad Nafsy API",
    "version": "2.0.0",
    "description": "Auto-generated from api/routes/*.js by tools/gen-openapi.mjs. Operation bodies are stubs; enrich summaries/schemas as needed."
  },
  "servers": [
    {
      "url": "https://mahadnafsy.com"
    },
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/admin/monitoring": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "GET /api/admin/monitoring",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/admin/payment-proofs": {
      "get": {
        "tags": [
          "payment-proofs"
        ],
        "summary": "GET /api/admin/payment-proofs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/admin/payment-proofs/{id}": {
      "patch": {
        "tags": [
          "payment-proofs"
        ],
        "summary": "PATCH /api/admin/payment-proofs/:id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/admin/payment-proofs/{id}/image": {
      "get": {
        "tags": [
          "payment-proofs"
        ],
        "summary": "GET /api/admin/payment-proofs/:id/image",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/me/payment-proof": {
      "post": {
        "tags": [
          "payment-proofs"
        ],
        "summary": "POST /api/me/payment-proof",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/me/payment-proofs": {
      "get": {
        "tags": [
          "payment-proofs"
        ],
        "summary": "GET /api/me/payment-proofs",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/me/progress": {
      "post": {
        "tags": [
          "lms"
        ],
        "summary": "POST /api/me/progress",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "lms"
        ],
        "summary": "GET /api/me/progress",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "payment-proofs"
        ],
        "summary": "PATCH /api/me/progress",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/openapi.json": {
      "get": {
        "tags": [
          "docs"
        ],
        "summary": "GET /api/openapi.json",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}